Restructure into common and schema_packages
