+++ pyproject.toml
@@ -136,7 +136,7 @@ where = ["src"]
 skip = [".github/*"]
 
 [project.entry-points.'nomad.plugin']
-schema = "nomad_auto_xrd:schema_entry_point"
+schema = "nomad_auto_xrd.schema_packages:schema_entry_point"
 models_app = "nomad_auto_xrd.apps:models_app_entry_point"
 example_upload = "nomad_auto_xrd.example_uploads:example_upload_entry_point"
 auto_xrd_training_action = "nomad_auto_xrd.actions.training:training_action"
