+++ src/nomad_auto_xrd/common/models.py
